+++ minga-web/package.json
@@ -9,6 +9,7 @@
     "axios": "^1.3.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.8.2",
     "react-scripts": "5.0.1",
     "sweetalert2": "^11.7.2",
     "web-vitals": "^2.1.4"
