+++ minga-web/src/components/Footer/Footer.jsx
