+++ minga-web/src/components/Footer/footer.css
