--- minga-web/src/components/MangaForm/mangaform.css
@@ -1,60 +1,0 @@
-* {
-    padding: 0;
-    margin: 0;
-    box-sizing: border-box;
-}
-h1 {
-    width:50vw;
-    display: flex;
-    font-size: 1.2rem;
-    text-align: center;
-}
-
-.content-form {
-    width: 100vw;
-    height: 100vh;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-fieldset {
-    border: none;
-    display: flex;
-    width: 50vw;
-    border-radius: 0;
-    margin: 0;
-    padding: 0;
-    background-color: transparent;
-
-}
-
-.inputMove {
-    border-radius: 0;
-    border-bottom: 1px solid black;
-}
-#selectMove {
-    border: 0;
-    width: 50vw;
-    color: gray;
-    font-size: 0.9rem;
-    background-color: transparent;
-    border-bottom: 1px solid black;
-}
-
-.btn-manga {
-    background: #4338CA;
-    border-radius: 3rem  ;
-    border: 1px solid #4338CA;
-    height: 2.5rem;
-    font-family: Roboto;
-    font-size: 24px;
-    font-weight: 700;
-    line-height: 28px;
-    letter-spacing: 0em;
-    text-align: center;
-    color: white;
-    font-size: 1.2rem;
-
-
-}
