+++ minga-web/src/layouts/MainLayout/MainLayout.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import NavBar from '../../components/Navbar/NavBar'
+import NavBar from '../../components/NavBar/NavBar'
 
 function MainLayout(props) {
     return (
